tests/wpt/web-platform-tests/css/css-multicol/parsing/column-span-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Multi-column Layout: getComputedValue().columnSpan</title>
<link rel="help" href="https://drafts.csswg.org/css-multicol/#propdef-column-span">
<meta name="assert" content="column-span computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("column-span", "none");
test_computed_value("column-span", "all");
</script>
</body>
</html>
